+++ packages/app/src/runner/SpecRunnerOpenMode.vue
@@ -9,7 +9,7 @@
   />
   <AdjustRunnerStyleDuringScreenshot
     id="main-pane"
-    class="flex border-gray-900"
+    class="flex"
   >
     <AutomationElement />
     <AutomationDisconnected