+++ packages/driver/cypress/e2e/commands/querying/querying.cy.js
@@ -82,6 +82,12 @@ describe('src/cy/commands/querying', () => {
       cy.get('doesNotExist')
     })
 
+    it('respects null withinSubject', () => {
+      cy.get('#list').within(() => {
+        cy.get('#upper', { withinSubject: null })
+      })
+    })
+
     describe('custom elements', () => {
       // <foobarbazquux>custom element</foobarbazquux>
 
@@ -1012,6 +1018,15 @@ describe('src/cy/commands/querying', () => {
         expect($span.length).to.eq(1)
         expect($span.get(0)).to.eq(span.get(0))
       })
+    })
+
+    // https://github.com/cypress-io/cypress/issues/25225
+    it('returns only one element when given multiple subjects directly match selector', () => {
+      // A case with only a text selector
+      cy.get('button').contains('submit').should('have.length', 1)
+
+      // A case with a filter + text selector
+      cy.get('div').contains('div', 'foo').should('have.length', 1)
     })
 
     // https://github.com/cypress-io/cypress/issues/25019