+++ packages/proxy/test/unit/http/request-middleware.spec.ts
@@ -167,13 +167,13 @@ describe('http/request-middleware', () => {
     })
 
     // CDP can determine whether or not the request is xhr | fetch, but the extension or electron cannot
-    it('provides resourceTypeAndCredentialManager with resourceType if able to determine from header (xhr)', async () => {
+    it('provides requestedWithAndCredentialManager with requestedWith if able to determine from header (xhr)', async () => {
       const ctx = {
         getAUTUrl: sinon.stub().returns('http://localhost:8080'),
         remoteStates: {
           isPrimarySuperDomainOrigin: sinon.stub().returns(false),
         },
-        resourceTypeAndCredentialManager: {
+        requestedWithAndCredentialManager: {
           get: sinon.stub().returns({}),
         },
         req: {
@@ -190,18 +190,18 @@ describe('http/request-middleware', () => {
 
       await testMiddleware([ExtractCypressMetadataHeaders], ctx)
       .then(() => {
-        expect(ctx.resourceTypeAndCredentialManager.get).to.have.been.calledWith('http://localhost:8080', `xhr`)
+        expect(ctx.requestedWithAndCredentialManager.get).to.have.been.calledWith('http://localhost:8080', `xhr`)
       })
     })
 
     // CDP can determine whether or not the request is xhr | fetch, but the extension or electron cannot
-    it('provides resourceTypeAndCredentialManager with resourceType if able to determine from header (fetch)', async () => {
+    it('provides requestedWithAndCredentialManager with requestedWith if able to determine from header (fetch)', async () => {
       const ctx = {
         getAUTUrl: sinon.stub().returns('http://localhost:8080'),
         remoteStates: {
           isPrimarySuperDomainOrigin: sinon.stub().returns(false),
         },
-        resourceTypeAndCredentialManager: {
+        requestedWithAndCredentialManager: {
           get: sinon.stub().returns({}),
         },
         req: {
@@ -218,19 +218,19 @@ describe('http/request-middleware', () => {
 
       await testMiddleware([ExtractCypressMetadataHeaders], ctx)
       .then(() => {
-        expect(ctx.resourceTypeAndCredentialManager.get).to.have.been.calledWith('http://localhost:8080', `fetch`)
+        expect(ctx.requestedWithAndCredentialManager.get).to.have.been.calledWith('http://localhost:8080', `fetch`)
       })
     })
 
-    it('sets the resourceType and credentialsLevel on the request from whatever is returned by resourceTypeAndCredentialManager if conditions apply', async () => {
+    it('sets the requestedWith and credentialsLevel on the request from whatever is returned by requestedWithAndCredentialManager if conditions apply', async () => {
       const ctx = {
         getAUTUrl: sinon.stub().returns('http://localhost:8080'),
         remoteStates: {
           isPrimarySuperDomainOrigin: sinon.stub().returns(false),
         },
-        resourceTypeAndCredentialManager: {
+        requestedWithAndCredentialManager: {
           get: sinon.stub().returns({
-            resourceType: 'fetch',
+            requestedWith: 'fetch',
             credentialStatus: 'same-origin',
           }),
         },