+++ system-tests/projects/svelte-webpack/cypress.config.js
@@ -2,6 +2,7 @@ const { defineConfig } = require('cypress')
 
 module.exports = defineConfig({
   component: {
+    experimentalSingleTabRunMode: true,
     devServer: {
       framework: 'svelte',
       bundler: 'webpack',