+++ tooling/v8-snapshot/src/setup/v8-snapshot-entry.ts
@@ -1,0 +1,1 @@
+import '@packages/server/v8-snapshot-entry'